requirements-api.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.0
alembic>=1.13.0
pydantic>=2.5.0
httpx>=0.26.0
requests>=2.31.0
anthropic>=0.18.0
redis>=5.0.0
kafka-python>=2.0.2
python-multipart>=0.0.6
python-dotenv>=1.0.0
boto3>=1.34.0
deltalake>=0.18.0
pandas>=2.0.0
pyarrow>=14.0.0
